+++ lib/essentials/conversation_graph/application/handles/handle_projection_repository.dart
@@ -2,10 +2,14 @@ class HandleProjectionResult {
   const HandleProjectionResult({
     required this.examinedHandleCount,
     required this.insertedHandleCount,
+    this.normalizedHandleCount = 0,
+    this.preservedUnnormalizedHandleCount = 0,
   });
 
   final int examinedHandleCount;
   final int insertedHandleCount;
+  final int normalizedHandleCount;
+  final int preservedUnnormalizedHandleCount;
 }
 
 class HandleIdentityProjectionResult {